backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
alembic==1.14.0
pymysql==1.1.1
psycopg[binary]==3.2.3
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
joblib==1.4.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.12
httpx==0.27.2
orjson==3.10.12
